backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1 